refactor: posts and comments, add:votes.go
