+++ internal/repository/postgres/posts.go
@@ -2,6 +2,7 @@ package postgres
 
 import (
 	"database/sql"
+	"log"
 	"time"
 
 	"github.com/Shalqarov/forum/internal/domain"
@@ -50,6 +51,24 @@ const (
 	INNER JOIN "user" AS u
 		ON p.user_id = u.user_id
 	ORDER BY "date" DESC`
+
+	queryGetVotedPostsByUserID = `
+	SELECT p."id", p."user_id", p."title", p."category", p."date", u.username
+	FROM "post" AS p
+	INNER JOIN "user" AS u
+		ON p.id = v.post_id
+	INNER JOIN "post_votes" AS v
+		ON v."user_id" = u.id AND v.post_id = p.id
+	WHERE u.id=$1 AND v."vote"=1
+	ORDER BY p."date" DESC
+	`
+
+	queryGetVotesCountByPostID = `
+	SELECT "vote", count("vote") 
+	FROM "post_votes"
+	WHERE post_id = $1 
+	GROUP BY "vote"
+	ORDER BY "vote" desc`
 )
 
 func (r *repo) CreatePost(post *domain.Post) (int64, error) {
@@ -114,6 +133,52 @@ func (u *repo) GetAllPosts() ([]*domain.PostDTO, error) {
 	}
 	defer rows.Close()
 	return scanPostDTORows(rows)
+}
+
+func (u *repo) GetVotedPostsByUserID(userID int64) ([]*domain.PostDTO, error) {
+	rows, err := u.db.Query(queryGetVotedPostsByUserID, userID)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+	posts := []*domain.PostDTO{}
+	for rows.Next() {
+		post := &domain.PostDTO{}
+		err := rows.Scan(&post.ID, &post.UserID, &post.Title, &post.Category, &post.CreatedAt, &post.Author)
+		if err != nil {
+			return nil, err
+		}
+		posts = append(posts, post)
+	}
+	return posts, nil
+}
+
+func (u *repo) GetVotesCountByPostID(postID int64) (*domain.Vote, error) {
+	rows, err := u.db.Query(queryGetVotesCountByPostID, postID)
+	if err != nil {
+		return nil, err
+	}
+	votes := &domain.Vote{
+		Like:    0,
+		Dislike: 0,
+	}
+	for rows.Next() {
+		var voteType int64
+		var cnt int64
+		err := rows.Scan(&voteType, &cnt)
+		if err != nil {
+			return nil, err
+		}
+		switch voteType {
+		case 1:
+			votes.Like = uint64(cnt)
+		case -1:
+			votes.Dislike = uint64(cnt)
+		default:
+			log.Println("Get Votes count bug:", voteType, cnt)
+		}
+	}
+	return votes, nil
 }
 
 func scanPostDTORows(rows *sql.Rows) ([]*domain.PostDTO, error) {
